+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/display/DisplayerSpec.kt
@@ -48,6 +48,7 @@ class DisplayerSpec {
             miniAppInfo = TEST_MA,
             miniAppMessageBridge = miniAppMessageBridge,
             miniAppNavigator = mock(),
+            miniAppFileChooser = mock(),
             miniAppCustomPermissionCache = mock(),
             downloadedManifestCache = mock(),
             queryParams = TEST_URL_PARAMS
@@ -57,6 +58,7 @@ class DisplayerSpec {
         appUrl = TEST_MA_URL,
         miniAppMessageBridge = miniAppMessageBridge,
         miniAppNavigator = mock(),
+        miniAppFileChooser = mock(),
         miniAppCustomPermissionCache = mock(),
         downloadedManifestCache = mock(),
         queryParams = TEST_URL_PARAMS