+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/display/RealMiniAppDisplaySpec.kt
@@ -20,6 +20,7 @@ import com.rakuten.tech.mobile.miniapp.js.MiniAppMessageBridge
 import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.test.runBlockingTest
 import org.amshove.kluent.*
+import org.junit.After
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -34,11 +35,12 @@ class RealMiniAppDisplaySpec {
     private lateinit var basePath: String
     private lateinit var realDisplay: RealMiniAppDisplay
     private lateinit var miniAppAnalytics: MiniAppAnalytics
+    private val activityScenario = ActivityScenario.launch(TestActivity::class.java)
     private val miniAppMessageBridge: MiniAppMessageBridge = mock()
 
     @Before
     fun setup() {
-        ActivityScenario.launch(TestActivity::class.java).onActivity { activity ->
+        activityScenario.onActivity { activity ->
             context = activity
             basePath = context.filesDir.path
             realDisplay = RealMiniAppDisplay(
@@ -46,6 +48,7 @@ class RealMiniAppDisplaySpec {
                 miniAppInfo = TEST_MA,
                 miniAppMessageBridge = miniAppMessageBridge,
                 miniAppNavigator = mock(),
+                miniAppFileChooser = mock(),
                 hostAppUserAgentInfo = TEST_HA_NAME,
                 miniAppCustomPermissionCache = mock(),
                 downloadedManifestCache = mock(),
@@ -57,12 +60,18 @@ class RealMiniAppDisplaySpec {
         }
     }
 
+    @After
+    fun finish() {
+        activityScenario.close()
+    }
+
     @Test
     fun `should pass MiniAppInfo forUrl through the constructor`() {
         val realDisplay = RealMiniAppDisplay(
             appUrl = "",
             miniAppMessageBridge = miniAppMessageBridge,
             miniAppNavigator = mock(),
+            miniAppFileChooser = mock(),
             hostAppUserAgentInfo = TEST_HA_NAME,
             miniAppCustomPermissionCache = mock(),
             downloadedManifestCache = mock(),