+++ miniapp/src/test/java/com/rakuten/tech/mobile/miniapp/js/userinfo/UserInfoBridgeSpec.kt
@@ -3,6 +3,7 @@ package com.rakuten.tech.mobile.miniapp.js.userinfo
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import com.google.gson.Gson
 import com.nhaarman.mockitokotlin2.*
+import com.nhaarman.mockitokotlin2.mock
 import com.rakuten.tech.mobile.miniapp.*
 import com.rakuten.tech.mobile.miniapp.TEST_CALLBACK_ID
 import com.rakuten.tech.mobile.miniapp.TEST_MA_DISPLAY_NAME
@@ -13,15 +14,16 @@ import com.rakuten.tech.mobile.miniapp.TEST_MA_VERSION_TAG
 import com.rakuten.tech.mobile.miniapp.TEST_USER_NAME
 import com.rakuten.tech.mobile.miniapp.display.WebViewListener
 import com.rakuten.tech.mobile.miniapp.js.*
+import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridge.Companion.ERR_ACCESS_TOKEN_NOT_MATCH_MANIFEST
+import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridge.Companion.ERR_ACCESS_TOKEN_NO_PERMISSION
 import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridge.Companion.ERR_GET_ACCESS_TOKEN
 import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridge.Companion.ERR_GET_CONTACTS
+import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridge.Companion.ERR_GET_CONTACTS_NO_PERMISSION
 import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridge.Companion.ERR_GET_PROFILE_PHOTO
 import com.rakuten.tech.mobile.miniapp.js.userinfo.UserInfoBridge.Companion.ERR_GET_USER_NAME
 import com.rakuten.tech.mobile.miniapp.permission.*
 import com.rakuten.tech.mobile.miniapp.storage.DownloadedManifestCache
-import org.amshove.kluent.When
-import org.amshove.kluent.calling
-import org.amshove.kluent.itReturns
+import org.amshove.kluent.*
 import org.junit.Before
 import org.junit.Test
 import org.junit.runner.RunWith
@@ -39,11 +41,6 @@ class UserInfoBridgeSpec {
     )
     private val profilePhotoCallbackObj = CallbackObj(
         action = ActionType.GET_PROFILE_PHOTO.action,
-        param = null,
-        id = TEST_CALLBACK_ID
-    )
-    private val tokenCallbackObj = CallbackObj(
-        action = ActionType.GET_ACCESS_TOKEN.action,
         param = null,
         id = TEST_CALLBACK_ID
     )
@@ -84,6 +81,10 @@ class UserInfoBridgeSpec {
         whenever(customPermissionCache.hasPermission(
             miniAppInfo.id, MiniAppCustomPermissionType.CONTACT_LIST)
         ).thenReturn(true)
+        whenever(customPermissionCache.hasPermission(
+            miniAppInfo.id, MiniAppCustomPermissionType.ACCESS_TOKEN)
+        ).thenReturn(true)
+        When calling downloadedManifestCache.getAccessTokenPermissions(TEST_MA_ID) itReturns TEST_ATP_LIST
     }
 
     /** start region: onGetUserName */
@@ -109,11 +110,11 @@ class UserInfoBridgeSpec {
     }
 
     private fun createUserInfoBridgeWrapper(userInfoBridgeDispatcher: UserInfoBridgeDispatcher): UserInfoBridge {
-        val userInfoBridgeWrapper =
-            UserInfoBridge()
+        val userInfoBridgeWrapper = UserInfoBridge()
         userInfoBridgeWrapper.setMiniAppComponents(
             bridgeExecutor,
             customPermissionCache,
+            downloadedManifestCache,
             TEST_MA.id
         )
         userInfoBridgeWrapper.setUserInfoBridgeDispatcher(userInfoBridgeDispatcher)
@@ -254,6 +255,12 @@ class UserInfoBridgeSpec {
 
     /** start region: access token */
     private val testToken = TokenData("test_token", 0)
+    private val tokenCallbackObj = CallbackObj(
+            action = ActionType.GET_ACCESS_TOKEN.action,
+            param = Gson().toJson(TEST_ATP1),
+            id = TEST_CALLBACK_ID
+    )
+
     private fun createAccessTokenImpl(
         hasAccessToken: Boolean,
         canGetToken: Boolean
@@ -262,6 +269,7 @@ class UserInfoBridgeSpec {
             object : UserInfoBridgeDispatcher {
                 override fun getAccessToken(
                     miniAppId: String,
+                    accessTokenScope: AccessTokenScope,
                     onSuccess: (tokenData: TokenData) -> Unit,
                     onError: (message: String) -> Unit
                 ) {
@@ -292,9 +300,83 @@ class UserInfoBridgeSpec {
         val userInfoBridgeDispatcher = Mockito.spy(createAccessTokenImpl(true, false))
         val userInfoBridgeWrapper = Mockito.spy(createUserInfoBridgeWrapper(userInfoBridgeDispatcher))
 
-        userInfoBridgeWrapper.onGetAccessToken(tokenCallbackObj.id)
+        userInfoBridgeWrapper.onGetAccessToken(tokenCallbackObj)
 
         verify(bridgeExecutor).postError(tokenCallbackObj.id, errMsg)
+    }
+
+    @Test
+    fun `postError should be called when access token permission hasn't been allowed`() {
+        val userInfoBridgeDispatcher = Mockito.spy(createAccessTokenImpl(true, true))
+        val userInfoBridgeWrapper = Mockito.spy(createUserInfoBridgeWrapper(userInfoBridgeDispatcher))
+
+        val errMsg = "$ERR_GET_ACCESS_TOKEN $ERR_ACCESS_TOKEN_NO_PERMISSION"
+        whenever(customPermissionCache.hasPermission(
+            miniAppInfo.id, MiniAppCustomPermissionType.ACCESS_TOKEN)
+        ).thenReturn(false)
+
+        userInfoBridgeWrapper.onGetAccessToken(tokenCallbackObj)
+
+        verify(bridgeExecutor).postError(tokenCallbackObj.id, errMsg)
+    }
+
+    @Test
+    fun `postError should be called when the requested audience and scope not match the define in manifest`() {
+        val userInfoBridgeDispatcher = Mockito.spy(createAccessTokenImpl(true, true))
+        val userInfoBridgeWrapper = Mockito.spy(createUserInfoBridgeWrapper(userInfoBridgeDispatcher))
+        val errMsg = "$ERR_GET_ACCESS_TOKEN $ERR_ACCESS_TOKEN_NOT_MATCH_MANIFEST"
+
+        val tokenCallbackObj = CallbackObj(
+            action = ActionType.GET_ACCESS_TOKEN.action,
+            param = null,
+            id = TEST_CALLBACK_ID
+        )
+        userInfoBridgeWrapper.onGetAccessToken(tokenCallbackObj)
+        verify(bridgeExecutor).postError(tokenCallbackObj.id, errMsg)
+
+        val atp2 = AccessTokenScope(audience = "aud", scopes = mutableListOf("scopeB"))
+        val tokenCallbackObj2 = CallbackObj(
+            action = ActionType.GET_ACCESS_TOKEN.action,
+            param = Gson().toJson(atp2),
+            id = TEST_CALLBACK_ID + '2'
+        )
+        userInfoBridgeWrapper.onGetAccessToken(tokenCallbackObj2)
+        verify(bridgeExecutor).postError(tokenCallbackObj2.id, errMsg)
+
+        val atp3 = AccessTokenScope(audience = "aud1", scopes = mutableListOf())
+        val tokenCallbackObj3 = CallbackObj(
+            action = ActionType.GET_ACCESS_TOKEN.action,
+            param = Gson().toJson(atp3),
+            id = TEST_CALLBACK_ID + '3'
+        )
+        userInfoBridgeWrapper.onGetAccessToken(tokenCallbackObj3)
+        verify(bridgeExecutor).postError(tokenCallbackObj3.id, errMsg)
+
+        val atp4 = AccessTokenScope(audience = "aud2", scopes = mutableListOf("scopeA", "scopeB"))
+        val tokenCallbackObj4 = CallbackObj(
+            action = ActionType.GET_ACCESS_TOKEN.action,
+            param = Gson().toJson(atp4),
+            id = TEST_CALLBACK_ID + '4'
+        )
+        userInfoBridgeWrapper.onGetAccessToken(tokenCallbackObj4)
+        verify(bridgeExecutor).postError(tokenCallbackObj4.id, errMsg)
+    }
+
+    @Test
+    fun `should return the audience & scopes for miniapp`() {
+        val userInfoBridgeDispatcher = Mockito.spy(createAccessTokenImpl(true, true))
+        val userInfoBridgeWrapper = Mockito.spy(createUserInfoBridgeWrapper(userInfoBridgeDispatcher))
+
+        val atp2 = AccessTokenScope(audience = "aud1", scopes = mutableListOf("scopeB"))
+        val tokenCallbackObj2 = CallbackObj(
+            action = ActionType.GET_ACCESS_TOKEN.action,
+            param = Gson().toJson(atp2),
+            id = TEST_CALLBACK_ID
+        )
+        userInfoBridgeWrapper.onGetAccessToken(tokenCallbackObj2)
+
+        verify(bridgeExecutor).postValue(tokenCallbackObj2.id, Gson().toJson(testToken))
+        testToken.scopes shouldEqual atp2
     }
 
     @Test
@@ -302,7 +384,24 @@ class UserInfoBridgeSpec {
         val userInfoBridgeDispatcher = Mockito.spy(createAccessTokenImpl(true, true))
         val userInfoBridgeWrapper = Mockito.spy(createUserInfoBridgeWrapper(userInfoBridgeDispatcher))
 
-        userInfoBridgeWrapper.onGetAccessToken(tokenCallbackObj.id)
+        userInfoBridgeWrapper.onGetAccessToken(tokenCallbackObj)
+        verify(bridgeExecutor).postValue(tokenCallbackObj.id, Gson().toJson(testToken))
+    }
+
+    @Test
+    fun `will call deprecated function to get token when still using old implementation`() {
+        val userInfoBridgeDispatcher = Mockito.spy(object : UserInfoBridgeDispatcher {
+            override fun getAccessToken(
+                miniAppId: String,
+                onSuccess: (tokenData: TokenData) -> Unit,
+                onError: (message: String) -> Unit
+            ) {
+                onSuccess.invoke(testToken)
+            }
+        })
+        val userInfoBridgeWrapper = Mockito.spy(createUserInfoBridgeWrapper(userInfoBridgeDispatcher))
+
+        userInfoBridgeWrapper.onGetAccessToken(tokenCallbackObj)
 
         verify(bridgeExecutor).postValue(tokenCallbackObj.id, Gson().toJson(testToken))
     }
@@ -346,7 +445,7 @@ class UserInfoBridgeSpec {
         val userInfoBridgeDispatcher = Mockito.spy(createContactsImpl(true, true))
         val userInfoBridgeWrapper = Mockito.spy(createUserInfoBridgeWrapper(userInfoBridgeDispatcher))
 
-        val errMsg = "$ERR_GET_CONTACTS Permission has not been accepted yet for getting contacts."
+        val errMsg = "$ERR_GET_CONTACTS $ERR_GET_CONTACTS_NO_PERMISSION"
         whenever(customPermissionCache.hasPermission(
             miniAppInfo.id, MiniAppCustomPermissionType.CONTACT_LIST)
         ).thenReturn(false)